Added the deactivate link

+++ app/Http/Controllers/UserController.php
@@ -68,7 +68,9 @@ class UserController extends Controller
         foreach ($users as $key => $value) {
             $id = md5($value->id);
             $passreset = url("user/passwordReset/$id");
-            $statusChange = url("user/status/$id");
+            $statusChange = "<a href='".url("user/status/$id/deactivate")."'>Deactivate</a>";
+            if (null !== $value->deleted_at)
+                $statusChange = "<a href='".url("user/status/$id/activate")."'>Activate</a>";
             $delete = url("user/delete/$id");
             $last_access = (null === $value->last_access) ? "" : date('M d, Y (H:i)', strtotime($value->last_access));
             $status = (null === $value->deleted_at) ? "<span class='label label-success'>Active</span>" : "<span class='label label-danger'>Inactive</span>";
@@ -80,7 +82,7 @@ class UserController extends Controller
             $row .= '<td>'.$value->user_type.'</td>';
             $row .= '<td>'.$last_access.'</td>';
             $row .= '<td>'.$status.'</td>';
-            $row .= '<td><a href="'.$passreset.'">Reset Password</a> | <a href="'.$statusChange.'">Deactivate</a> | <a href="'.$delete.'">Delete</a></td>';
+            $row .= '<td><a href="'.$passreset.'">Reset Password</a> | '.$statusChange.' | <a href="'.$delete.'">Delete</a></td>';
             $row .= '</tr>';
         }
 
@@ -237,10 +239,15 @@ class UserController extends Controller
         }
     }
 
-    public function updateState(Request $request, $user = null)
+    public function updateState(Request $request, $user, $status)
     {
         $user = self::__unHashUser($user);
-        $user->delete();
+        $status = strtolower($status);
+        if ($status == 'deactivate') {
+            $user->delete();
+        } else if ($status == 'activate') {
+            $user->restore();
+        }
         return back();
     }
